+++ OctoAwesome/OctoAwesome/IUserInterfaceExtension.cs
@@ -1,0 +1,16 @@
+﻿using OctoAwesome.Entities;
+
+namespace OctoAwesome
+{
+    /// <summary>
+    /// Interface for UI dependend <see cref="EntityComponent"/>
+    /// </summary>
+    public interface IUserInterfaceExtension
+    {
+        /// <summary>
+        /// Register the extending UI.
+        /// </summary>
+        /// <param name="manager">UI-Manager</param>
+        void Register(IUserInterfaceManager manager);
+    }
+}
